requirements.txt
flask==3.0.0
scikit-learn==1.3.2
numpy==1.26.2
mlflow==2.9.2
dvc==3.35.0
pyyaml==6.0.1
pytest==7.4.3
pytest-cov==4.1.0
pylint==3.0.3
black==23.12.1
python-dotenv==1.0.0
requests==2.31.0